Move lookup_country_in_pfxnummult_array()
